fix: outparsing errored - i assumed that the output format of the julia program is always the same. But this is not the case. In the case of MOFs, it gives also the clustering as a pdb file. The new implementation tries to check for that. In the longer run, we should look closer how this is done in Julia and then adapt our parser. - started adding dependencies for the caching

+++ crystal_analysis_webservice/core.py
@@ -64,8 +64,12 @@ def run_topology_analysis(fileContent: str, extension: str = "cif"):
 
                 raise ValueError(errorline)
             if out_:
-                out = out_[-2]
-                os.unlink(out_[0].split()[-1])
+                if len(out_) > 2:
+                    out = out_[-2]
+                    os.unlink(out_[0].split()[-1])
+                else:
+                    out = out_[0]
+
             else:
                 raise ValueError
